Erreur au niveau affichage de transaction a valider au niveau de use et le telechargement des transactions

+++ lib/services/partenariat/transaction_partenariat_service.dart
@@ -246,7 +246,7 @@ class TransactionPartenaritModel {
       produit: json['produit'],
       quantite: json['quantite'] is int
           ? json['quantite']
-          : int.parse(json['quantite'].toString()),
+          : double.parse(json['quantite'].toString()).toInt(),
       prixUnitaire: json['prixUnitaire'] is double
           ? json['prixUnitaire']
           : json['prix_unitaire'] is double
